add range_search() to IndexRefine (#4022)
Summary:
This is very convenient to have `range_seach()` in `IndexRefine`. Unlike the plain `search()` method, `range_search()` just reevaluates the computed distances from the baseline index. The labels are not re-sorted according to new distances, because this is not listed as a requirement in a method description
https://github.com/facebookresearch/faiss/blob/4cd5226a10817abd81be879da9092eb73bb337a6/faiss/Index.h#L150-L161
https://github.com/facebookresearch/faiss/blob/4cd5226a10817abd81be879da9092eb73bb337a6/faiss/impl/AuxIndexStructures.h#L35

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4022

Reviewed By: mnorris11

Differential Revision: D66116082

Pulled By: gtwang01

fbshipit-source-id: 915aca2570d5863c876c9497d4c885e270b9b220