fix `IVFPQFastScan::RangeSearch()` on the `ARM` architecture (#4247)
Summary:
the problem happens if `radius - normalizers[2 * q + 1]` is negative. Thus, it is possible to provide reasonable parameters to `IVFPQFastScan::RangeSearch()` and get an empty result.

I have no idea WHY (hardware implementation, it seems), but the following code
```C++
#include <cstddef>
#include <cstdint>
#include <iostream>

int main() {
    float f = -25.5f;
    uint16_t t = f;
    std::cout << t << std::endl;
    return 0;
}
```
prints `65511` on `x86` and `0` on ARM on the same compiler.

Thus, it is needed to wrap the `float` value with `int` to preserve a correct cast:
```C++
    uint16_t t = (int)f;
```

Who would have thought...

It is useful to find some C++ compiler command line flags that will generate a compilation error on such a behavior.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4247

Reviewed By: junjieqi, satymish

Differential Revision: D71427185

Pulled By: gtwang01

fbshipit-source-id: 3ff3a9d3bb523e48bb9512c380c042bb1c2decdb