Fix building on mingw (#4420)
Summary:
This restores the ability to compile with mingw introduced in https://github.com/facebookresearch/faiss/issues/4145. I am not sure why msvc does allow this, but using the `_FMT` macro without any `vaargs` causes a trailing comma in the arguments list. which is not allowed in c++. This switches it over to the `_MSG` macro instead avoiding this issue.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4420

Reviewed By: pankajsingh88

Differential Revision: D77827581

Pulled By: ramilbakhshyiev

fbshipit-source-id: f63564546511e88084dbacc71458448d3c403de9