BinaryHNSWCagra Struct; Allow base_level_only (#4478)
Summary:
This PR adds a new struct -- IndexBinaryHNSWCagra, that is used for the interoperability with GpuIndexBinaryCagra. It also adds the serialization, deserialization functions for this new struct and resolves a seg fault in the interop. Furthermore, this new struct allows for only the base layer to be built and searched in the hnsw graph (similar to IndexHNSWCagra).

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4478

Reviewed By: gtwang01

Differential Revision: D79365863

Pulled By: mnorris11

fbshipit-source-id: a93559b84ccbe297f7193466e05fcfe0ebf82783