Remove invalid assertion checking #neighbors == graph degree (#4528)
Summary:
# Description
As fixed in https://github.com/facebookresearch/faiss/pull/4516, graph degree can be adjusted when small number of vectors are provided. Therefore, the number of neighbors (precisely, the maximum number of neighbors) can be smaller than graph degree after it was adjusted.

This PR is to remove the assertion check that fails if #neighbors not equals to graph degree, which is not always true.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4528

Reviewed By: junjieqi

Differential Revision: D80057041

Pulled By: mnorris11

fbshipit-source-id: 128f902fe346cd7c0c4c94664d5d166a5d4c4b16