Use Development.Module component in CMake FindPython (#4549)
Summary:
This PR updates the CMake configuration for the Python extension.

The current configuration specifies `Development` python component that includes the unnecessary `Development.Embed` sub-component. For a Python extension, only `Development.Module` component is required. The current configuration leads to a cmake build failure in the `manylinux` environment.

https://cmake.org/cmake/help/latest/module/FindPython.html

**Additional context**
I found this issue while I was testing the `scikit-build-core` backend to build wheel packages in this repository:
https://github.com/kyamagu/faiss-build

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4549

Reviewed By: junjieqi

Differential Revision: D80952883

Pulled By: limqiying

fbshipit-source-id: 793bc369fae8e9f1745f91d7f573f77494a1b886