bugfix: add a macro guard for avx512 operation (#4539)
Summary:
### What to fix
There is a small bug in the faiss/utils/rabitq_simd.h that some avx512 only supported operation is not guarded by a corresponding macro. As a result, the compile will fail if the arch does not support avx512.

### What I did
I add the macro guard.

### What I checked
I have compiled locally.

This should be an easy fix.

Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4539

Reviewed By: mnorris11

Differential Revision: D80734590

Pulled By: junjieqi

fbshipit-source-id: 603b94f29931576330eee1a8d1dcb73164df47f2