Allow unaligned fast-scan (#4623)
Summary:
Pull Request resolved: https://github.com/facebookresearch/faiss/pull/4623

Codes can be loaded from disk, so alignment is not guaranteed. This diff allows unaligned pointer in pq fast scan code.

I checked what the overhead was in low level instructions in godbolt. All 3 approaches are in P1997520050.
- No conditional
- Template parameter conditional (same number of instructions as no conditional, so it is probably not slower?)
- New parameter (increases number of instructions, making it slower)

Tried microbenchmarking in a loop. Seems to be the same between unaligned and aligned.

Unit Test
-
For non avx512, the code handles it correctly with the unit test. P1998522307 shows the error without any `c.loadu(codes)` handling.

Do we need to update avx512 version?
-
- The unit test passes with whatever code size I tried. I might have to change other things like nsq? May need input here.

Reviewed By: junjieqi

Differential Revision: D84364872

fbshipit-source-id: 59bae66d97d292d1096899280c94bb87ad15c69e